refactor: creating separate file for runblockframe
